requirements.txt
folium==0.14.0
Pillow==10.0.1
pytest==7.4.2